# API-Go/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  feature.CreateFeatureRequest:
    properties:
      action_level:
        enum:
        - READ
        - WRITE
        - DELETE
        - PUBLISH
        - APPROVE
        type: string
      code:
        maxLength: 60
        type: string
      description:
        type: string
      module_id:
        type: string
      name:
        maxLength: 120
        type: string
    required:
    - action_level
    - code
    - module_id
    - name
    type: object
  feature.CreateModuleRequest:
    properties:
      code:
        maxLength: 30
        type: string
      description:
        type: string
      display_order:
        type: integer
      icon_key:
        type: string
      name:
        maxLength: 100
        type: string
    required:
    - code
    - name
    type: object
  feature.FeatureResponse:
    properties:
      action_level:
        type: string
      code:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      module_id:
        type: string
      name:
        type: string
    type: object
  feature.ListFeaturesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/feature.FeatureResponse'
        type: array
      page:
        type: integer
      total:
        type: integer
    type: object
  feature.ListModulesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/feature.ModuleResponse'
        type: array
      page:
        type: integer
      total:
        type: integer
    type: object
  feature.ModuleResponse:
    properties:
      code:
        type: string
      display_order:
        type: integer
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
    type: object
  feature.UpdateFeatureRequest:
    properties:
      action_level:
        enum:
        - READ
        - WRITE
        - DELETE
        - PUBLISH
        - APPROVE
        type: string
      description:
        type: string
      is_active:
        type: boolean
      name:
        type: string
    type: object
  feature.UpdateModuleRequest:
    properties:
      description:
        type: string
      display_order:
        type: integer
      icon_key:
        type: string
      is_active:
        type: boolean
      name:
        type: string
    type: object
  response.Body:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 'API REST de Identity and Access Management (IAM): autenticación JWT,
    RBAC, usuarios, roles, funcionalidades, sesiones y auditoría.'
  title: API JeussAirel (IAM)
  version: "1.0"
paths:
...
  /features:
    get:
      parameters:
      - description: Número de página (default 1)
        in: query
        name: page
        type: integer
      - description: Tamaño de página (1-100, default 20)
        in: query
        name: page_size
        type: integer
      responses:
        "200":
          description: Funcionalidades obtenidas
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/feature.ListFeaturesResponse'
              type: object
        "403":
          description: Acceso denegado (IDENTITY_ROLE_MANAGE:write)
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Listar funcionalidades
      tags:
      - API
    post:
      parameters:
      - description: Datos de la funcionalidad
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/feature.CreateFeatureRequest'
      responses:
        "201":
          description: Funcionalidad creada
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/feature.FeatureResponse'
              type: object
        "400":
          description: Datos inválidos
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Acceso denegado (IDENTITY_ROLE_MANAGE:write)
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Crear funcionalidad
      tags:
      - API
  /features/{id}:
    delete:
      parameters:
      - description: UUID de la funcionalidad
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Funcionalidad eliminada
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Acceso denegado (IDENTITY_ROLE_MANAGE:write)
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Funcionalidad no encontrada
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Eliminar funcionalidad
      tags:
      - API
    get:
      parameters:
      - description: UUID de la funcionalidad
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Funcionalidad encontrada
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/feature.FeatureResponse'
              type: object
        "403":
          description: Acceso denegado (IDENTITY_ROLE_MANAGE:write)
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Funcionalidad no encontrada
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Obtener funcionalidad por ID
      tags:
      - API
    put:
      parameters:
      - description: UUID de la funcionalidad
        in: path
        name: id
        required: true
        type: string
      - description: Campos a actualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/feature.UpdateFeatureRequest'
      responses:
        "200":
          description: Funcionalidad actualizada
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/feature.FeatureResponse'
              type: object
        "403":
          description: Acceso denegado (IDENTITY_ROLE_MANAGE:write)
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Funcionalidad no encontrada
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Actualizar funcionalidad
      tags:
      - API
  /modules:
    get:
      parameters:
      - description: Número de página (default 1)
        in: query
        name: page
        type: integer
      - description: Tamaño de página (1-100, default 20)
        in: query
        name: page_size
        type: integer
      responses:
        "200":
          description: Módulos obtenidos
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/feature.ListModulesResponse'
              type: object
        "403":
          description: Acceso denegado (IDENTITY_ROLE_MANAGE:write)
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Listar módulos
      tags:
      - API
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos del módulo
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/feature.CreateModuleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Módulo creado
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/feature.ModuleResponse'
              type: object
        "400":
          description: Datos inválidos
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Sin permisos (IDENTITY_ROLE_MANAGE:write)
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Crear módulo
      tags:
      - Modules
  /modules/{id}:
    delete:
      parameters:
      - description: UUID del módulo
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Módulo eliminado
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Acceso denegado (IDENTITY_ROLE_MANAGE:write)
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Módulo no encontrado
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Eliminar módulo
      tags:
      - API
    get:
      parameters:
      - description: UUID del módulo
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Módulo encontrado
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/feature.ModuleResponse'
              type: object
        "403":
          description: Acceso denegado (IDENTITY_ROLE_MANAGE:write)
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Módulo no encontrado
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Obtener módulo por ID
      tags:
      - Modules
    put:
      parameters:
      - description: UUID del módulo
        in: path
        name: id
        required: true
        type: string
      - description: Campos a actualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/feature.UpdateModuleRequest'
      responses:
        "200":
          description: Módulo actualizado
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/feature.ModuleResponse'
              type: object
        "403":
          description: Acceso denegado (IDENTITY_ROLE_MANAGE:write)
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Módulo no encontrado
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Actualizar módulo
  /modules/{id}/features:
    get:
      parameters:
      - description: UUID del módulo
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Funcionalidades del módulo
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/feature.ListFeaturesResponse'
              type: object
        "403":
          description: Acceso denegado (IDENTITY_ROLE_MANAGE:write)
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Módulo no encontrado
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Listar funcionalidades de un módulo
      tags:
      - API
...
securityDefinitions:
  BearerAuth:
    description: 'Token JWT de acceso. Uso: `Authorization: Bearer <access_token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"